fix: switch project
